Add files via upload
This project covered a variety of topics, including setting up a custom network, installing Ghost on Docker, and setting up WordPress on Ubuntu. I discovered that Nginx is used for reverse proxying and serving web content. The article offered detailed instructions for setting up Nginx, starting it, and adding reverse proxy logs for the Ghost site. Additionally, the project covered the installation and testing of Ghost on Docker as well as the LAMP stack installation of WordPress on Ubuntu. The project was informative overall and gave a good overview of each of the subjects covered.